# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.FollowedListResponse:
    properties:
      followed:
        items:
          $ref: '#/definitions/response.UserDTO'
        type: array
      user_id:
        type: integer
      user_name:
        type: string
    type: object
  response.FollowersCountResponse:
    properties:
      followers_count:
        type: integer
      user_id:
        type: integer
      user_name:
        type: string
    type: object
  response.FollowersListResponse:
    properties:
      followers:
        items:
          $ref: '#/definitions/response.UserDTO'
        type: array
      user_id:
        type: integer
      user_name:
        type: string
    type: object
...
  response.UserDTO:
    properties:
      user_id:
        type: integer
      user_name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: suporte@socialmeli.com
    name: Suporte API
  description: API REST para rede social do Mercado Livre
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SocialMeli API
  version: "1.0"
paths:
...
  /users/{id}/follow/{userIdToFollow}:
    post:
      consumes:
      - application/json
      description: Permite que um usuário siga um vendedor (US-0001)
      parameters:
      - description: ID do usuário que vai seguir
        in: path
        name: id
        required: true
        type: integer
      - description: ID do usuário a ser seguido
        in: path
        name: userIdToFollow
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Seguir um vendedor
      tags:
      - users-follow
  /users/{id}/followed/list:
    get:
      consumes:
      - application/json
      description: Retorna a lista de vendedores que um usuário segue (US-0004, US-0008)
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      - default: name_asc
        description: 'Ordenação: name_asc ou name_desc'
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.FollowedListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar seguidos
      tags:
      - Follow
  /users/{id}/followers/count:
    get:
      consumes:
      - application/json
      description: Retorna o número de seguidores de um vendedor (US-0002)
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.FollowersCountResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obter contagem de seguidores
      tags:
      - users-followers
  /users/{id}/followers/list:
    get:
      consumes:
      - application/json
      description: Retorna a lista de usuários que seguem um vendedor (US-0003, US-0008)
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      - default: name_asc
        description: 'Ordenação: name_asc ou name_desc'
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.FollowersListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar seguidores
      tags:
      - users-followers
  /users/{id}/unfollow/{userIdToFollow}:
    post:
      consumes:
      - application/json
      description: Remove o vínculo de follow entre usuários (US-0007)
      parameters:
      - description: ID do usuário que vai deixar de seguir
        in: path
        name: id
        required: true
        type: integer
      - description: ID do usuário a ser deixado de seguir
        in: path
        name: userIdToFollow
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deixar de seguir um vendedor
      tags:
      - users-follow
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Digite "Bearer" seguido de um espaço e depois o token JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"